# public/openapi.yaml
openapi: 3.0.3
info:
  title: allaboard.dev API
  version: "1.0"
  description: REST API for the Allaboard board-climbing community platform.
servers:
  - url: https://www.allaboard.dev
    description: Production (www)
  - url: https://allaboard.dev
    description: Production (apex)
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: allaboard_session
      description: iron-session encrypted cookie set after Google OAuth login.
    apiKeyAuth:
      type: apiKey
      in: query
      name: token
      description: Personal API token from your profile page. Pass as ?token=<value>.
paths:
...
  /api/climbs/{id}/ticks:
    get:
      summary: List all ticks for a climb, newest first.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of tick objects, each embedding the submitting user's handle, display name, avatar color, and profile picture URL.
        "500":
          description: database error.
    post:
      summary: Submit a tick (send record) for a climb.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: The created tick object with status 201.
        "400":
          description: rating is missing or outside 1–4.
        "401":
          description: not authenticated.
        "404":
          description: the climb does not exist.
        "500":
          description: database error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rating
              properties:
                rating:
                  type: string
                  description: star rating 1–4.
                date:
                  type: string
                  description: ISO date string ("YYYY-MM-DD"); defaults to today.
                sent:
                  type: string
                  description: whether the climb was completed.
                attempts:
                  type: string
                  description: number of attempts.
                suggestedGrade:
                  type: string
                  description: the climber's grade opinion ("V0"–"V18").
                comment:
                  type: string
                  description: free-form send notes.
                instagramUrl:
                  type: string
                  description: URL to an Instagram post/reel of the send.
      security:
        - cookieAuth: []
...
  /api/comments/{id}:
    patch:
      summary: Update a comment's body.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The updated comment.
        "400":
          description: body is empty.
        "401":
          description: not authenticated.
        "403":
          description: not the comment owner.
        "404":
          description: comment does not exist.
        "500":
          description: database error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
              properties:
                body:
                  type: string
                  description: new comment text.
      security:
        - apiKeyAuth: []
        - cookieAuth: []
    delete:
      summary: Delete a comment.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content on success.
        "401":
          description: not authenticated.
        "403":
          description: not the comment owner.
        "404":
          description: comment does not exist.
        "500":
          description: database error.
      security:
        - apiKeyAuth: []
        - cookieAuth: []
  /api/comments:
    get:
      summary: List all comments for a tick, ordered oldest-first, with replies nested.
      parameters:
        - name: tickId
          in: query
          required: true
          description: UUID of the tick to fetch comments for.
          schema:
            type: string
      responses:
        "200":
          description: Array of top-level Comment objects, each with a replies array.
        "400":
          description: tickId is missing.
        "500":
          description: database error.
    post:
      summary: Post a comment on a tick.
      responses:
        "201":
          description: The created Comment object with status 201.
        "400":
          description: tickId or body are missing.
        "401":
          description: not authenticated.
        "404":
          description: the tick does not exist.
        "500":
          description: database error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tickId
                - body
              properties:
                tickId:
                  type: string
                  description: UUID of the tick being commented on.
                body:
                  type: string
                  description: comment text.
                parentCommentId:
                  type: string
                  description: UUID of the parent comment when replying.
      security:
        - apiKeyAuth: []
        - cookieAuth: []
...